requirements.txt
openai==1.69.0
python_box==7.3.2
